requirements.txt
streamlit
google-generativeai
python-dotenv
tiktoken
openai
wheel
reportlab